ujenziiq-backend/requirements.txt
Django==4.2.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.2.0
djoser==2.2.0
django-filter==23.2
Pillow==10.0.0
python-dotenv==1.0.0
